SPAM_PHISING_DETECTOR/requirements.txt
nltk
scikit-learn
gunicorn
langchain
langchain_core
langchain_google_genai
langgraph
typedict
pydantic
python-dotenv
fastapi
uvicorn